Making Possible changes

+++ src/landing_page/home/HomePage.js
@@ -8,10 +8,26 @@ import Education from './Education.js';
 import OpenAccount from '../OpenAccount';
 import Footer from '../Footer.js';
 import { useEffect } from 'react';
+import  axios from 'axios';
 function HomePage() {
     useEffect(() => {
-        localStorage.clear();
-        console.log("All localStorage cleared on first load");
+        const user = localStorage.getItem('user');
+
+        if (user) {
+            // Send logout request to backend
+            axios.post('http://localhost:3002/logout')
+                .then(() => {
+                    console.log("User logged out from backend.");
+                })
+                .catch(err => {
+                    console.error("Logout error:", err);
+                });
+
+            // Clear frontend session
+            localStorage.clear();
+            console.log("LocalStorage cleared on HomePage load.");
+
+        }
     }, []);
     return (
         <>
